+++ react/components/Footer.js
