+++ react/components/ShelfItem.js
